+++ example/src/app/formly-example/formly-input-examples.component.ts
@@ -1,0 +1,36 @@
+import { Component } from '@angular/core';
+import { FormGroup } from "@angular/forms";
+import { FormlyFieldConfig } from "@ngx-formly/core";
+
+@Component({
+  selector: 'app-formly-input-examples',
+  template: `
+    <form [formGroup]="form" (ngSubmit)="onSubmit(model)">
+      <formly-form [form]="form" [fields]="fields" [model]="model"></formly-form>
+      <button type="submit" class="btn btn-default">Submit</button>
+    </form>
+  `,
+  styles: [
+  ]
+})
+export class FormlyInputExamplesComponent {
+
+  form = new FormGroup({});
+  model = { email: 'email@gmail.com' };
+  fields: FormlyFieldConfig[] = [
+    {
+      key: 'email',
+      type: 'input',
+      templateOptions: {
+        label: 'Email address',
+        placeholder: 'Enter email',
+        required: true,
+      }
+    }
+  ];
+
+  onSubmit(model) {
+    console.log(model);
+  }
+
+}